[APP-873] refactor: move video related files to a base module: aptoide-ui
